*migrate wasm test app from CRA to CRN *modify the renameExports script to not rename wasm exports for `wasm.` prefixes.
