--- examples/wasm-test-app/.eslintignore
@@ -1,3 +1,0 @@
-.eslintrc.js
-src/pkg/ezkl.js
-src/pkg/snippets/
