--- examples/wasm-test-app/src/index.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import TestScript from './WASMTests';
-import reportWebVitals from './reportWebVitals';
-
-const rootComponent = process.env.REACT_APP_ENTRY_POINT === 'index_test.tsx' ? <TestScript /> : <App />;
-
-const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
-);
-root.render(
-  <React.StrictMode>
-    {rootComponent}
-  </React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
