+++ package.json
@@ -15,6 +15,9 @@
   "author": "Jason Morton",
   "license": "Apache-2.0",
   "devDependencies": {
+    "@nomicfoundation/hardhat-toolbox": "^3.0.0",
+    "@typechain/hardhat": "^8.0.0",
+    "@types/jest": "^27.5.2",
     "@types/json-bigint": "^1.0.1",
     "@types/node": "^20.4.5",
     "@typescript-eslint/eslint-plugin": "^6.2.0",
@@ -22,13 +25,11 @@
     "eslint": "^8.46.0",
     "eslint-config-prettier": "^8.9.0",
     "eslint-plugin-react": "^7.33.1",
-    "typescript": "^5.1.6",
-    "@nomicfoundation/hardhat-toolbox": "^3.0.0",
-    "@typechain/hardhat": "^8.0.0",
-    "@types/jest": "^27.5.2",
     "hardhat": "^2.16.1",
     "jest": "^29.6.1",
-    "ts-jest": "^29.1.1"
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.1.6",
+    "create-next-app": "^13.4.12"
   },
   "dependencies": {
     "ethers": "^6.6.7",
